Restructures CheckoutSuccessPage
