+++ src/components/Checkout/CheckoutSuccess/Information/Information.styled.jsx
@@ -1,0 +1,5 @@
+import styled from "styled-components";
+
+export const Info = styled.div`
+  flex-grow: 1;
+`;
