+++ src/components/Checkout/CheckoutSuccess/Information/index.jsx
@@ -1,0 +1,1 @@
+export { default } from "./Information";
