+++ src/components/Checkout/CheckoutSuccess/OrderSummary/OrderSummary.styled.jsx
@@ -1,22 +1,5 @@
 import styled from "styled-components";
 
-export const Container = styled.div`
-  max-width: 1000px;
-  margin-inline: auto;
-`;
-
-export const Heading = styled.div`
-  display: flex;
-`;
-
-export const CheckoutContainer = styled.div`
-  @media (min-width: 700px) {
-    display: flex;
-    justify-content: center;
-    gap: 5rem;
-  }
-`;
-
 export const List = styled.div`
   flex-grow: 1;
   > h4 {
@@ -76,13 +59,4 @@ export const CartItem = styled.li`
       margin-bottom: 0;
     }
   }
-`;
-
-export const Info = styled.div`
-  flex-grow: 1;
-`;
-
-export const LinkContainer = styled.div`
-  text-align: center;
-  margin-top: 3rem;
 `;
