+++ src/components/Checkout/CheckoutSuccess/OrderSummary/index.jsx
@@ -1,0 +1,1 @@
+export { default } from "./OrderSummary";
