+++ src/pages/CheckoutSuccessPage/CheckoutSuccessPage.jsx
@@ -1,0 +1,8 @@
+import CheckoutSuccess from "../../components/Checkout/CheckoutSuccess";
+
+const CheckoutSuccessPage = () => {
+  document.title = "Successful Order | Shop";
+  return <CheckoutSuccess />;
+};
+
+export default CheckoutSuccessPage;
